chore: rename slash-menu-popover
